First commit: An animated JQuery thermometer that displays both positive and negative values
Based originally off of:
http://jsfiddle.net/GeekyJohn/vQ4Xn/

This is an animated jQuery thermometer that will display temperature
values for you based either upon DOM values or input arguments.
Originally, this was a one-way positive-values-only type thermometer to
display fundraising amounts, (e.g. $12,562 out of a $50,000 goal), but
I've re-tooled it to be a useful temperature thermometer for both
positive and negative values.

I've commented out the track increments since they don't tend to form a
useful measurement for the thermometer, but left them in place in case
you think otherwise.

As the "0" marker line tends to be offset slightly due to the line
itself being the top border of the mid-marker's div, I've provided a
"minAdjust" variable in the code you can use to nudge the line up or
down slightly.

HTML5 Compliant.